+++ packages/utils/src/index.ts
@@ -34,6 +34,6 @@
 export { deepMerge } from './deep-merge'
 export { streamToString, stringToStream, objectToStream, streamToObject } from './stream'
 export { z, isObjectRecord } from './zod'
-export type { Z, ZodType, ZodError, input, output } from './zod'
+export type { Z, Zod } from './zod'
 export { parseJson, parseObjectFromJson, parseArrayFromJson } from './json'
 export { parseYaml } from './yaml'